authorization implemented to all controllers

+++ Backend/src/main/java/cz/cvut/fit/genepi/presentationLayer/controller/card/InvasiveTestEegController.java
@@ -7,6 +7,7 @@ import cz.cvut.fit.genepi.businessLayer.service.UserService;
 import cz.cvut.fit.genepi.businessLayer.service.card.InvasiveTestEegService;
 import cz.cvut.fit.genepi.dataLayer.entity.PatientEntity;
 import cz.cvut.fit.genepi.dataLayer.entity.card.InvasiveTestEegEntity;
+import cz.cvut.fit.genepi.util.AuthorizationChecker;
 import cz.cvut.fit.genepi.util.TimeConverter;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.security.core.Authentication;
@@ -20,6 +21,7 @@ import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
 
+import javax.servlet.http.HttpServletRequest;
 import javax.validation.Valid;
 import java.util.List;
 import java.util.Locale;
@@ -43,8 +45,10 @@ public class InvasiveTestEegController {
 
     @RequestMapping(value = "/patient/{patientId}/invasive-test-eeg/create", method = RequestMethod.GET)
     public String invasiveTestEegCreateGET(
-            @PathVariable("patientId") Integer patientId, Locale locale, Model model) {
-
+            @PathVariable("patientId") Integer patientId, Locale locale, Model model, HttpServletRequest request) {
+        if (!AuthorizationChecker.checkAuthoritaion(request)) {
+            return "deniedView";
+        }
         model.addAttribute("patient", patientService.getPatientDisplayByIdWithDoctor(patientId));
         model.addAttribute("invasiveTestEeg", new InvasiveTestEegVO());
         return "patient/invasiveTestEeg/formView";
@@ -54,8 +58,10 @@ public class InvasiveTestEegController {
     public String complicationEditGET(
             @PathVariable("patientId") Integer patientId,
             @PathVariable("invasiveTestEegId") Integer invasiveTestEegId,
-            Locale locale, Model model) {
-
+            Locale locale, Model model, HttpServletRequest request) {
+        if (!AuthorizationChecker.checkAuthoritaion(request)) {
+            return "deniedView";
+        }
         model.addAttribute("patient", patientService.getPatientDisplayByIdWithDoctor(patientId));
         model.addAttribute("invasiveTestEeg", invasiveTestEegService.getById(InvasiveTestEegVO.class, InvasiveTestEegEntity.class, invasiveTestEegId));
         return "patient/invasiveTestEeg/formView";
@@ -63,16 +69,19 @@ public class InvasiveTestEegController {
 
     /**
      * Adds the invasiveTestEEG.
-          * @param result          the result
+     *
+     * @param result the result
      * @return the string
      */
     @RequestMapping(value = "/patient/{patientId}/invasive-test-eeg/save", method = RequestMethod.POST)
     public String invasiveTestEegSavePOST(
             @ModelAttribute("invasiveTestEeg") @Valid InvasiveTestEegVO invasiveTestEeg, BindingResult result,
             @PathVariable("patientId") Integer patientId,
-            Locale locale, Model model) {
-
-        if (result.hasErrors()|| TimeConverter.compareDates(patientService.getPatientByIdWithDoctor(patientId).getBirthday(), invasiveTestEeg.getDate())) {
+            Locale locale, Model model, HttpServletRequest request) {
+        if (!AuthorizationChecker.checkAuthoritaion(request)) {
+            return "deniedView";
+        }
+        if (result.hasErrors() || TimeConverter.compareDates(patientService.getPatientByIdWithDoctor(patientId).getBirthday(), invasiveTestEeg.getDate())) {
             model.addAttribute("patient", patientService.getPatientDisplayByIdWithDoctor(patientId));
             return "patient/invasiveTestEeg/formView";
         } else {
@@ -87,7 +96,7 @@ public class InvasiveTestEegController {
                 }
             }
             if (!isSuperdoctor)
-                patientService.findByID(PatientEntity.class,patientId).setVerified(false);
+                patientService.findByID(PatientEntity.class, patientId).setVerified(false);
             invasiveTestEeg.setPatientId(patientId);
             invasiveTestEegService.save(InvasiveTestEegEntity.class, invasiveTestEeg);
             return "redirect:/patient/" + patientId + "/invasive-test-eeg/list";
@@ -98,8 +107,10 @@ public class InvasiveTestEegController {
     public String invasiveTestEegDeleteGET(
             @PathVariable("patientId") Integer patientId,
             @PathVariable("invasiveTestEegId") Integer invasiveTestEegId,
-            Locale locale, Model model) {
-
+            Locale locale, Model model, HttpServletRequest request) {
+        if (!AuthorizationChecker.checkAuthoritaion(request)) {
+            return "deniedView";
+        }
         invasiveTestEegService.delete(InvasiveTestEegEntity.class, invasiveTestEegId);
         return "redirect:/patient/" + patientId + "/invasive-test-eeg/list";
     }
@@ -107,17 +118,19 @@ public class InvasiveTestEegController {
     /**
      * Handles the GET request to hide invasiveTestEeg.
      *
-     * @param patientId   the id of a patient whom we are creating an invasiveTestEeg.
-     * @param locale      the user's locale.
-     * @param model       the model to be filled for view.
+     * @param patientId the id of a patient whom we are creating an invasiveTestEeg.
+     * @param locale    the user's locale.
+     * @param model     the model to be filled for view.
      * @return the address to which the user will be redirected.
      */
     @RequestMapping(value = "/patient/{patientId}/invasive-test-eeg/{invasiveTestEegId}/hide", method = RequestMethod.GET)
     public String invasiveTestEegHideGET(
             @PathVariable("patientId") Integer patientId,
             @PathVariable("invasiveTestEegId") Integer invasiveTestEegId,
-            Locale locale, Model model) {
-
+            Locale locale, Model model, HttpServletRequest request) {
+        if (!AuthorizationChecker.checkAuthoritaion(request)) {
+            return "deniedView";
+        }
         invasiveTestEegService.hide(invasiveTestEegId);
         return "redirect:/patient/" + patientId + "/invasive-test-eeg/list";
     }
@@ -125,17 +138,19 @@ public class InvasiveTestEegController {
     /**
      * Handles the GET request to unhide invasiveTestEeg.
      *
-     * @param patientId   the id of a patient whom we are creating an invasiveTestEeg.
-     * @param locale      the user's locale.
-     * @param model       the model to be filled for view.
+     * @param patientId the id of a patient whom we are creating an invasiveTestEeg.
+     * @param locale    the user's locale.
+     * @param model     the model to be filled for view.
      * @return the address to which the user will be redirected.
      */
     @RequestMapping(value = "/patient/{patientId}/invasive-test-eeg/{invasiveTestEegId}/unhide", method = RequestMethod.GET)
     public String invasiveTestEegUnhideGET(
             @PathVariable("patientId") Integer patientId,
             @PathVariable("invasiveTestEegId") Integer invasiveTestEegId,
-            Locale locale, Model model) {
-
+            Locale locale, Model model, HttpServletRequest request) {
+        if (!AuthorizationChecker.checkAuthoritaion(request)) {
+            return "deniedView";
+        }
         invasiveTestEegService.unhide(invasiveTestEegId);
         // TODO: address to get back to admin module where is list od hidden
         // records.
@@ -151,7 +166,10 @@ public class InvasiveTestEegController {
 
     @RequestMapping(value = "/patient/{patientId}/invasive-test-eeg/list", method = RequestMethod.GET)
     public String invasiveTestEegListGET(
-            @PathVariable("patientId") Integer patientId, Locale locale, Model model) {
+            @PathVariable("patientId") Integer patientId, Locale locale, Model model, HttpServletRequest request) {
+        if (!AuthorizationChecker.checkAuthoritaion(request)) {
+            return "deniedView";
+        }
         PatientDisplayVO patient = patientService.getPatientDisplayByIdWithInvasiveTestEegList(patientId);
         model.addAttribute("beginningEpilepsy", TimeConverter.getAgeAtTheBeginningOfEpilepsy(patient));
         model.addAttribute("currentAge", TimeConverter.getCurrentAge(patient));
